refactor: extract proposal transformation logic into a shared utility and update status calculation logic
